add the dynamic operation support for initialize allowed curser
